+++ HermesMobile/HermesMobileApp.swift
@@ -69,6 +69,7 @@ struct HermesMobileApp: App {
 
     init() {
         BGTask.register()
+        MainThreadWatchdog.shared.start()
     }
 
     var body: some Scene {